requirements.txt
# Основной фреймворк
flet>=0.25.0

# Базы данных
chromadb>=0.4.0
aiosqlite>=0.19.0

# ИИ и обработка текста
ollama>=0.3.0
langchain>=0.1.0
langchain-community>=0.0.20
sentence-transformers>=2.2.0
openai>=1.0.0

# Обработка документов
pypdf2>=3.0.0
python-docx>=1.1.0
beautifulsoup4>=4.12.0
markdown2>=2.4.0

# Уведомления
python-telegram-bot>=20.0
pync>=2.0.3

# Утилиты
aiofiles>=23.0.0
python-dateutil>=2.8.0
pydantic>=2.0.0
loguru>=0.7.0
pyyaml>=6.0
